readline: prefer 5 (fast) over 6 (slow)
Running racket in a terminal with readline 6 has excruciating delays
in response and bizarrely slow pasting. The culprit is probably a bug
in readline v6.2 that causes the rl_event_hook to be called repeatedly
even when input is ready. See also the discussion here:
http://lists.gnu.org/archive/html/bug-readline/2012-06/msg00005.html

It's not clear whether racket can work around the issue in readline 6,
but it can at least use readline 5 instead when it is available.

original commit: ae843f9a6288000f63a25c06392229f2eebfae57