adapt to CS where callbacks are always atomic
While a `readline` is in progress, we'd like Racket threads to still
run. That was previously accomplished by setting a character-reading
(or byte-reading) callbacks that uses `read-char` or `read-byte` ---
but CS wants callbacks to be atomic, and so that doesn't work.

One possibility is to make the callback non-atomic anyway by using
`end-atomic` and `start-atomic`. That works in CS as long as no other
use of callbacks in a Racket place tries the same trick.

Another approach is to run `readline` in a separate OS thread, and
have a callback there bounce back to a Racket thread. That's some
bookkeeping and using some Chez Scheme-level facilities, but it works
out.

This commit tries the latter approach.